# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: LedgerTwo API
  version: 1.3.0-rc
  description: |
    LedgerTwo v1.3 当前实现正式契约。实际路径为 /api/...；
    /api/v1 alias 尚未实现，不能作为已上线接口使用。
    所有账本内业务接口必须使用 X-Ledger-Id 或账本 path 参数建立显式上下文，
    不再支持首账本 fallback。金额字段均为整数分。
servers:
  - url: /api
    description: Current compatibility API root
security: []
...
paths:
...
  /init/setup:
    post:
      tags: [init]
      summary: Initialize application, first user and first ledger
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InitSetupRequest"
      responses:
        "200":
          $ref: "#/components/responses/Success"
        "400":
          $ref: "#/components/responses/ValidationError"
        "409":
          $ref: "#/components/responses/Conflict"
  /auth/login:
    post:
      tags: [auth]
      summary: Login with username and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login succeeded and token cookie is set
          headers:
            Set-Cookie:
              schema:
                type: string
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /auth/me:
    get:
      tags: [auth]
      summary: Get current user
      security:
        - cookieAuth: []
      responses:
        "200":
          $ref: "#/components/responses/Success"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /ledgers/:
    get:
      tags: [ledgers]
      summary: List ledgers visible to current user
      description: 默认只返回 active；结果不承担当前账本 fallback。
      security:
        - cookieAuth: []
      parameters:
        - name: status
          in: query
          required: false
          schema:
            type: string
            enum: [active, archived, all]
            default: active
      responses:
        "200":
          $ref: "#/components/responses/Success"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      tags: [ledgers]
      summary: Create an active ledger
      description: 创建者成为唯一 Owner；创建同名账本合法。
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateLedgerRequest"
      responses:
        "201":
          $ref: "#/components/responses/Success"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /ledgers/{ledgerId}:
    parameters:
      - $ref: "#/components/parameters/LedgerIdPath"
    get:
      tags: [ledgers]
      summary: Get one active or archived ledger
      security:
        - cookieAuth: []
      responses:
        "200":
          $ref: "#/components/responses/Success"
        "403":
          $ref: "#/components/responses/Forbidden"
    patch:
      tags: [ledgers]
      summary: Rename an active ledger
      description: Owner only. Repeated stale versions return LEDGER_VERSION_CONFLICT.
      security:
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/IfMatchLedger"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
      responses:
        "200":
          $ref: "#/components/responses/Success"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
...
  /ledgers/{ledgerId}/archive:
    parameters:
      - $ref: "#/components/parameters/LedgerIdPath"
    post:
      tags: [ledgers]
      summary: Archive an active ledger
      description: Owner only. Archive is read-only and never creates settlement or rewrites history.
      security:
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/IfMatchLedger"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                acknowledge_unsettled_balance:
                  type: boolean
                  default: false
      responses:
        "200":
          $ref: "#/components/responses/Success"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
  /ledgers/{ledgerId}/restore:
    parameters:
      - $ref: "#/components/parameters/LedgerIdPath"
    post:
      tags: [ledgers]
      summary: Restore an archived ledger
      description: Owner only. This is the only archived-ledger mutation.
      security:
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/IfMatchLedger"
      responses:
        "200":
          $ref: "#/components/responses/Success"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
...
  /categories:
    get:
      tags: [metadata]
      summary: List categories
      security:
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/LedgerIdHeaderRequired"
        - name: include_archived
          in: query
          required: false
          schema:
            type: boolean
          description: When true, archived categories are included for historical transaction display.
      responses:
        "200":
          $ref: "#/components/responses/Success"
  /accounts:
    get:
      tags: [metadata]
      summary: List accounts
      security:
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/LedgerIdHeaderRequired"
      responses:
        "200":
          $ref: "#/components/responses/Success"
  /metadata/{kind}/:
    parameters:
      - $ref: "#/components/parameters/MetadataKindPath"
      - $ref: "#/components/parameters/LedgerIdHeaderRequired"
    get:
      tags: [metadata]
      summary: List metadata items
      security:
        - cookieAuth: []
      parameters:
        - name: include_archived
          in: query
          required: false
          schema:
            type: boolean
            default: false
      responses:
        "200":
          $ref: "#/components/responses/Success"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      tags: [metadata]
      summary: Create metadata item
      description: Owner only. kind supports categories, tags and accounts.
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MetadataUpsertRequest"
      responses:
        "201":
          $ref: "#/components/responses/Success"
        "400":
          $ref: "#/components/responses/ValidationError"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/Conflict"
  /metadata/{kind}/reorder:
    parameters:
      - $ref: "#/components/parameters/MetadataKindPath"
      - $ref: "#/components/parameters/LedgerIdHeaderRequired"
    post:
      tags: [metadata]
      summary: Reorder metadata items
      description: Owner only. Updates sort_order for the submitted ids; archived items may be omitted.
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MetadataReorderRequest"
      responses:
        "200":
          $ref: "#/components/responses/Success"
        "400":
          $ref: "#/components/responses/ValidationError"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /metadata/{kind}/{id}:
    parameters:
      - $ref: "#/components/parameters/MetadataKindPath"
      - $ref: "#/components/parameters/GenericIdPath"
      - $ref: "#/components/parameters/LedgerIdHeaderRequired"
    patch:
      tags: [metadata]
      summary: Update metadata item
      description: Owner only.
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MetadataUpsertRequest"
      responses:
        "200":
          $ref: "#/components/responses/Success"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
  /metadata/{kind}/{id}/archive:
    post:
      tags: [metadata]
      summary: Archive metadata item
      description: Owner only. Archived items are hidden from create/edit selectors but remain available for historical records.
      security:
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/MetadataKindPath"
        - $ref: "#/components/parameters/GenericIdPath"
        - $ref: "#/components/parameters/LedgerIdHeaderRequired"
      responses:
        "200":
          $ref: "#/components/responses/Success"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /metadata/{kind}/{id}/restore:
    post:
      tags: [metadata]
      summary: Restore archived metadata item
      description: Owner only.
      security:
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/MetadataKindPath"
        - $ref: "#/components/parameters/GenericIdPath"
        - $ref: "#/components/parameters/LedgerIdHeaderRequired"
      responses:
        "200":
          $ref: "#/components/responses/Success"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /transaction-defaults:
    get:
      tags: [transactions]
      summary: Get current user's quick transaction defaults
      description: Returns per-ledger, per-user transaction defaults. Archived category, account and tags are omitted from the response.
      security:
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/LedgerIdHeaderRequired"
      responses:
        "200":
          $ref: "#/components/responses/Success"
        "403":
          $ref: "#/components/responses/Forbidden"
  /transactions/:
    get:
      tags: [transactions]
      summary: List transactions
      security:
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/LedgerIdHeaderRequired"
        - $ref: "#/components/parameters/MonthQuery"
        - $ref: "#/components/parameters/TypeQuery"
      responses:
        "200":
          $ref: "#/components/responses/Success"
    post:
      tags: [transactions]
      summary: Create ordinary transaction
      security:
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/LedgerIdHeaderRequired"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTransactionRequest"
      responses:
        "200":
          $ref: "#/components/responses/Success"
        "400":
          $ref: "#/components/responses/ValidationError"
        "403":
          $ref: "#/components/responses/Forbidden"
  /transactions/{transactionId}:
    parameters:
      - $ref: "#/components/parameters/TransactionIdPath"
      - $ref: "#/components/parameters/LedgerIdHeaderRequired"
    get:
      tags: [transactions]
      summary: Get transaction detail
      security:
        - cookieAuth: []
      responses:
        "200":
          $ref: "#/components/responses/Success"
        "404":
          $ref: "#/components/responses/NotFound"
    patch:
      tags: [transactions]
      summary: Update transaction
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTransactionRequest"
      responses:
        "200":
          $ref: "#/components/responses/Success"
        "400":
          $ref: "#/components/responses/ValidationError"
        "403":
          $ref: "#/components/responses/Forbidden"
    delete:
      tags: [transactions]
      summary: Soft delete transaction
      security:
        - cookieAuth: []
      responses:
        "200":
          $ref: "#/components/responses/Success"
        "403":
          $ref: "#/components/responses/Forbidden"
  /transactions/batch-tag:
    post:
      tags: [transactions]
      summary: Batch update transaction tags
      security:
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/LedgerIdHeaderRequired"
      responses:
        "200":
          $ref: "#/components/responses/Success"
  /transactions/import/parse:
    post:
      tags: [import]
      summary: Parse CSV import file
      security:
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/LedgerIdHeaderRequired"
      responses:
        "200":
          $ref: "#/components/responses/Success"
        "400":
          $ref: "#/components/responses/ValidationError"
  /transactions/import/analyze:
    post:
      tags: [import]
      summary: Analyze import items
      security:
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/LedgerIdHeaderRequired"
      responses:
        "200":
          $ref: "#/components/responses/Success"
  /transactions/import/commit:
    post:
      tags: [import]
      summary: Commit import batch
      security:
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/LedgerIdHeaderRequired"
      responses:
        "200":
          $ref: "#/components/responses/Success"
  /imports/preview:
    post:
      tags: [import]
      summary: Create a CSV or WeChat XLSX preview batch
      description: Owner only. Preview does not create formal transactions.
      security:
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/LedgerIdHeaderRequired"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [source_type, file]
              properties:
                source_type:
                  type: string
                  enum: [wechat, alipay, generic]
                file:
                  type: string
                  format: binary
      responses:
        "201":
          description: Persisted import preview batch with classification snapshots.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportPreviewBatchResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
  /imports/{batchID}:
    get:
      tags: [import]
      summary: Get a preview batch and rows
      security:
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/LedgerIdHeaderRequired"
        - $ref: "#/components/parameters/BatchIdPath"
      responses:
        "200":
          description: Persisted import preview batch with server-recomputed summaries.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportPreviewBatchResponse"
        "404":
          $ref: "#/components/responses/NotFound"
  /imports/{batchID}/rows/{rowID}:
    patch:
      tags: [import]
      summary: Adjust one preview row
      security:
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/LedgerIdHeaderRequired"
        - $ref: "#/components/parameters/BatchIdPath"
        - $ref: "#/components/parameters/RowIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
      responses:
        "200":
          description: Updated import preview batch.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImportPreviewBatchResponse"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /imports/{batchID}/commit:
    post:
      tags: [import]
      summary: Commit a ready preview batch atomically
      security:
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/LedgerIdHeaderRequired"
        - $ref: "#/components/parameters/BatchIdPath"
      responses:
        "200":
          $ref: "#/components/responses/Success"
        "409":
          $ref: "#/components/responses/Conflict"
...
  /import-rules/{ruleID}/archive:
    post:
      tags: [import]
      summary: Archive an import suggestion rule
      security:
        - cookieAuth: []
      parameters:
        - $ref: "#/components/parameters/RuleIdPath"
        - $ref: "#/components/parameters/LedgerIdHeaderRequired"
      responses:
        "200":
          $ref: "#/components/responses/Success"
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: token
...
  responses:
...
    Success:
      description: Successful response
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/SuccessResponse"
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    Conflict:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    InitSetupRequest:
      type: object
      required:
        - ledger_name
        - user_a_username
        - user_a_display_name
        - user_a_password
        - user_b_username
        - user_b_display_name
        - user_b_password
      properties:
        ledger_name:
          type: string
        default_currency:
          type: string
          default: CNY
        user_a_username:
          type: string
        user_a_display_name:
          type: string
        user_a_password:
          type: string
          format: password
        user_b_username:
          type: string
        user_b_display_name:
          type: string
        user_b_password:
          type: string
          format: password
...